third_party/blink/web_tests/external/wpt/dom/nodes/Comment-in-doctype-expected.txt

This is a testharness.js-based test.
[FAIL] XML: Comment in doctype internal subset
  assert_equals: expected DocumentType node but got Comment node <!--x-->
Harness: the test ran to completion.
